Two-phase commit for iprop log and HDB
We used to update the iprop log and HDB in different orders depending on
the kadm5 operation, which then led to various race conditions.

The sequence of updates is now: iprop log, HDB, mark last log entry
committed.  Every kadm5 write operation replays transactions not yet
confirmed as committed.

The log now starts with a nop entry whose payload records the offset of
the last confirmed committed transaction.  This is kown as the "ubber
entry".

The iprop-log(1) command's dump and last-version sub-commands now take
an option, -n, for not locking the iprop log.  This allows admins to
observe the log even while it is locked, and will be used for testing to
detect when a slave has caught up with the master.