Two-phase commit for iprop log and HDB
We used to update the iprop log and HDB in different orders depending on
the kadm5 operation, which then led to various race conditions.

The sequence of updates is now: iprop log, HDB, mark last log entry
committed.  Every kadm5 write operation replays transactions not yet
confirmed as committed.

The log now starts with a nop entry whose payload records the offset of
the last confirmed committed transaction.  This is kown as the "ubber
entry".

The iprop log format and the protocol remain backwards-compatible with
earlier versions of Heimdal.  This is NOT a flag-day; there is NO need
to update all the slaves at once with the master, though it is advisable
in general.