Build Windows post "Make the KDC use a multi-process model."
The commit 352a7e94a34c768b7b0f356afdcfdf0f31eefc91 broke the Windows
build because it makes use of POSIX functionality not available on Windows
and because it fails to make use of the roken socket abstractions.

This patchset applies the roken abstractions and restores the prior
start_kdc() (formerly known as loop()) functionality on Windows so the
tree will compile.

Change-Id: I0da10e3b83c83a9e39aec2032b4665c00b2ffa0f