ipropd-master: also check dump mtime
With this commit we now identify an iprop "epoch" by the version number
and timestamp of the first entry in the log (the first past the ubber
entry).  New kadm5_log functions allow the caller to find this identity.
The ipropd-master daemon now uses this to determine if a dump is valid.

The idea is that if the log gets reset to version 1 when the current
version was numerically small, then thins should be OK anyways because
the times of the new entries will be later than the times of any old
entries with the same version numbers.