ipropd-slave: refactor for better error handling
This commit begins the process of hardening ipropd-slave against dying
at the slightest provocation.  In particular, the process of writing
incremental transactions from the master to the local log file is
refactored out of receive_loop() and into a (new) append_to_log_file()
which tries hard to recover from errors.

ipropd-slave should now recover from EFBIG, EDQUOT, and ENOSPC whenever
removing the log file and starting from scratch would help.

XXX No, we need to preserve the version number in the log, so we should
not unlink() nor ftruncate(), but kadm5_log_truncate(), and
kadm5_log_truncate() should learn how to unlink() to recover in some
cases, and it should not do it on the master.