krb5: Fix MEMORY ccache missing unlock
In mcc_store_cred() we were missing an unlock on ENOMEM.

We still have thread-safety issues in the iterators, where cred removal
and ccache destruction can stop iteration in the middle.  Also,
mcc_get_cache_next() really needs to take a lock on the current cache
before taking a lock on the next cache, but we need to make sure that
doesn't lead to deadlock.